% Generated by roxygen2 (4.0.2): do not edit by hand
\name{r.map.addGeoJSON}
\alias{r.map.addGeoJSON}
\title{r.map.addGeoJSON}
\usage{
r.map.addGeoJSON(m, latitude, longitude, popup = NULL, radius = NULL,
  fillColor = NULL, fillOpacity = NULL, color = NULL, opacity = NULL,
  weight = NULL, constantpopup = NULL, constantradius = NULL,
  constantfillColor = NULL, constantfillOpacity = NULL,
  constantcolor = NULL, constantopacity = NULL, constantweight = NULL)
}
\description{
r.map.addGeoJSON
}

