% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rleaflet.R
\name{r.map.new}
\alias{r.map.new}
\title{r.map.new}
\usage{
r.map.new(mapCenter = c(41.39, 2.16), mapZoom = 13,
  mapTiles = "MapQuestOpen.OSM", enablePopover = FALSE, fullScreen = TRUE)
}
\description{
r.map.new
}

